css 템플릿 적용 마무리

+++ src/main/webapp/admin/notice/deleteNoticeAction.jsp
@@ -5,10 +5,18 @@
 	// Controller
 	request.setCharacterEncoding("utf-8");
 
+	// 삭제할 공지의 정보가 넘어오지않았을때
+	if(request.getParameter("noticeNo") == null || request.getParameter("noticeNo").equals("")) { // 비로그인시
+		out.println("<script>alert('오류'); location.href='" + request.getContextPath() + "/admin/notice/noticeOne.jsp" + "';</script>");
+		return;
+	}
+	
 	int noticeNo = Integer.parseInt(request.getParameter("noticeNo"));
 	
 	// Model
 	NoticeDao noticeDao = new NoticeDao();
+	
+	// 공지 삭제 메소드
 	int row = noticeDao.deleteNotice(noticeNo);
 	
 	if(row == 1) {
@@ -17,7 +25,7 @@
 		return;
 	} else {
 		System.out.println("삭제실패");
-		response.sendRedirect(request.getContextPath()+"/admin/notice/noticeOne.jsp?noticeNo=" + noticeNo);
+		out.println("<script>alert('삭제에 실패했습니다'); location.href='" + request.getContextPath() + "/admin/notice/noticeOne.jsp?noticeNo=" + noticeNo + "';</script>");
 		return;
 	}
 %>
